+++ test/e2e/fixtures/start-files-watch.js
@@ -1,5 +1,5 @@
 const path = require("path");
-const { startServer } = require("../utils");
+const { startServer } = require("../support/utils");
 
 startServer(path.resolve(__dirname, "files-watch"), {
   watch: true,