+++ test/e2e/plugin-options.spec.js
@@ -8,7 +8,14 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 const path = require("path");
-const { startServer, stopServer, request, CliRunner, wait, fixturesFolder } = require("./utils");
+const {
+  startServer,
+  stopServer,
+  request,
+  CliRunner,
+  wait,
+  fixturesFolder,
+} = require("./support/utils");
 
 describe("plugin options", () => {
   let server;
@@ -21,8 +28,8 @@ describe("plugin options", () => {
       });
     });
 
-    afterAll(() => {
-      stopServer(server);
+    afterAll(async () => {
+      await stopServer(server);
     });
 
     it("should disable deprecated behaviors api path", async () => {