+++ CHANGELOG.md
@@ -10,7 +10,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 ### Added
+- feat(#78): Add alerts router
+- chore(lint): Add eslint plugin to avoid only in tests
 ### Changed
+- chore(deps): Update dependencies
+- test(e2e): Move utils to a support folder
+- feat: Add/remove routers on start/stop plugin methods, not in init method.
 ### Fixed
 ### Removed